Move IHealth interfaces to Mods.Common
