+++ OpenRA.Mods.Common/Graphics/SelectionBarsAnnotationRenderable.cs
@@ -10,6 +10,7 @@
 #endregion
 
 using OpenRA.Graphics;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
 using OpenRA.Traits;
