+++ OpenRA.Mods.Common/TraitsInterfaces.cs
@@ -679,4 +679,20 @@ namespace OpenRA.Mods.Common.Traits
 			}
 		}
 	}
+
+	public interface IHealthInfo : ITraitInfo
+	{
+		int MaxHP { get; }
+	}
+
+	public interface IHealth : IActorDeath
+	{
+		DamageState DamageState { get; }
+		int HP { get; }
+		int MaxHP { get; }
+		int DisplayHP { get; }
+
+		void InflictDamage(Actor self, Actor attacker, Damage damage, bool ignoreModifiers);
+		void Kill(Actor self, Actor attacker, BitSet<DamageType> damageTypes);
+	}
 }
